fix: SDK consistency - fix critical bugs, add missing methods, fix docs
Agent-Logs-Url: https://github.com/MathiasGonzalez/UruFacturaSDK/sessions/33bd7685-d961-490e-a223-f99a8bbd59c7

+++ src/UruFacturaSDK/Xml/CfeXmlBuilder.cs
@@ -91,7 +91,7 @@ public class CfeXmlBuilder
         WriteElement(w, "MntBruto", "1");
         if (cfe.Moneda != Moneda.PesoUruguayo)
         {
-            WriteElement(w, "TipoMoneda", cfe.Moneda.ToString());
+            WriteElement(w, "TipoMoneda", ((int)cfe.Moneda).ToString());
             if (cfe.TipoCambio.HasValue)
                 WriteElement(w, "TpoCambio", cfe.TipoCambio.Value.ToString("F4"));
         }
